+++ src/components/Form.tsx
@@ -1,8 +1,8 @@
-import React from 'react'
+import React from 'preact/compat'
 
-export default ({ handleSubmit, requestPath, setRequestPath }): JSX.Element => (
+export default ({ handleSubmit, requestPath, setRequestPath }): React.JSX.Element => (
   <form id='add-file' onSubmit={handleSubmit}>
-    <label htmlFor='inputContent' className='f5 ma0 pb2 aqua fw4 db'>CID (and path)</label>
+    <label htmlFor='inputContent' className='f5 ma0 pb2 aqua fw4 db'>CID, Content Path, or URL</label>
     <input
       className='input-reset bn black-80 bg-white pa3 w-100 mb3'
       id='inputContent'
@@ -10,7 +10,7 @@ export default ({ handleSubmit, requestPath, setRequestPath }): JSX.Element => (
       type='text'
       placeholder='/ipfs/bafk.../path/to/file'
       required
-      value={requestPath} onChange={(e) => setRequestPath(e.target.value)}
+      value={requestPath} onChange={(e) => setRequestPath(e.currentTarget.value)}
     />
   </form>
 )